fix(routine): SUM 반환 타입 long + caloriesBurned 컬럼 기본값 설정
- JPQL SUM on int field returns Long: sumCaloriesBurnedByMemberIdAndDate 반환 타입 int → long
- ddl-auto=update 시 NOT NULL 컬럼 추가 시 오류 방지: @Column(columnDefinition = "integer default 0") 설정

+++ backend/src/main/java/com/ssafy/manager/routine/infrastructure/persistence/RoutineSessionRepository.java
@@ -12,5 +12,5 @@ public interface RoutineSessionRepository extends JpaRepository<RoutineSession, 
     List<RoutineSession> findTop4ByRoutineIdOrderBySessionDateDesc(Long routineId);
 
     @Query("SELECT COALESCE(SUM(rs.caloriesBurned), 0) FROM RoutineSession rs WHERE rs.memberId = :memberId AND rs.sessionDate = :date")
-    int sumCaloriesBurnedByMemberIdAndDate(@Param("memberId") Long memberId, @Param("date") LocalDate date);
+    long sumCaloriesBurnedByMemberIdAndDate(@Param("memberId") Long memberId, @Param("date") LocalDate date);
 }
